+++ src/model/user.ts
@@ -17,15 +17,5 @@ const User = sequelize.define('User', {
   },
 });
 
-function addUser({ id, name }: UserColumns) {
-  return User.create({
-    id,
-    name,
-  });
-}
-
-async function getUser(id: string) {
-  return User.findByPk(id);
-}
-
-export { addUser, getUser };
+export {User};
+export type { UserColumns };
